// app/components/FounderMessage/type.ts
export type FounderMessageProps = {
    banner: string;
    bannerAlt: string;
    metaTitle: string;
    metaDescription: string;
    pageTitle: string;
    firstSection: {
        title: string;
        description: string;
        image: string;
        imageAlt: string;
        name: string;
        designation: string;
    }
}